requirement.txt
aioredis==1.3.1
asgiref==3.3.4
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.0.5
django-filter==2.4.0
django-import-export==2.5.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
et-xmlfile==1.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.1
incremental==21.3.0
Markdown==3.3.4
MarkupPy==1.14
msgpack==1.0.2
numpy==1.20.2
odfpy==1.4.1
openpyxl==3.0.7
pandas==1.2.4
Pillow==8.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
tablib==3.0.0
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
zope.interface==5.4.0
